fix issue # 8
